gravitic-nebula/pyproject.toml
[project]
name = "gravitic-nebula"
version = "0.1.0"
description = "Alternative Data Microservice for Gravitic Financial Platform"
requires-python = ">=3.9"
dependencies = [
    "pydantic",
    "rapidfuzz",              # Fast Fuzzy Matching (SOTA for blocking)
    "sentence-transformers",  # Vector Embeddings (SOTA for semantics)
    "scikit-learn",          # Cosine Similarity / Utils
    "pandas",
    "chromadb",              # Vector Store for Entity Maps
    "firecrawl-py",          # SOTA Scrapping
    "google-generativeai",   # SOTA LLM (Gemini)
    "python-dotenv",         # Environment management
    "sqlalchemy",            # SOTA ORM for persistence
    "trafilatura",           # SOTA content extraction & cleaning
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
